+++ nym-vpn-core/crates/nym-exclude/src/main.rs
@@ -1,0 +1,14 @@
+// Copyright 2016-2026 Mullvad VPN AB. All Rights Reserved.
+// Copyright 2026 Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+#[cfg(target_os = "linux")]
+mod linux;
+
+fn main() {
+    #[cfg(target_os = "linux")]
+    linux::main();
+
+    #[cfg(not(target_os = "linux"))]
+    println!("This program is only supposed to run on Linux.")
+}