+++ nym-vpn-core/crates/nym-firewall/src/lib.rs
@@ -10,6 +10,9 @@ use ipnetwork::Ipv6Network;
 #[cfg(not(target_os = "android"))]
 use nym_dns::ResolvedDnsConfig;
 
+#[cfg(target_os = "linux")]
+use nym_cgroup::v2::CGroup2;
+
 #[cfg(target_os = "macos")]
 #[path = "macos.rs"]
 mod imp;
@@ -31,7 +34,6 @@ mod imp;
 mod imp;
 
 mod net;
-mod split_tunnel;
 pub use net::{
     AllowedClients, AllowedEndpoint, AllowedTunnelTraffic, Endpoint, TransportProtocol,
     TunnelInterface, TunnelMetadata,
@@ -395,10 +397,21 @@ pub struct FirewallArguments {
     pub initial_state: InitialFirewallState,
     /// This argument is required for the blocked state to configure the firewall correctly.
     pub allow_lan: bool,
-    /// Specifies the firewall mark used to identify traffic that is allowed to be excluded from
-    /// the tunnel and _leaked_ during blocked states.
+    /// Firewall mark is used to mark traffic which should be able to bypass the tunnel
     #[cfg(target_os = "linux")]
     pub fwmark: u32,
+    /// The table ID will be used for the routing table that will route all traffic through the
+    /// tunnel interface.
+    #[cfg(target_os = "linux")]
+    pub table_id: u32,
+    /// The cgroup2 used for split tunneling.
+    /// Traffic from processes in this cgroup2 should be allowed outside the tunnel.
+    #[cfg(target_os = "linux")]
+    pub excluded_cgroup2: Option<CGroup2>,
+    /// The net_cls id of the v1 cgroup used for split tunneling.
+    /// This is used as a fallback to [`Self::excluded_cgroup2`] since old kernels don't support cgroups v2.
+    #[cfg(target_os = "linux")]
+    pub net_cls: Option<u32>,
 }
 
 /// State to enter during firewall init.
@@ -417,16 +430,6 @@ impl Firewall {
         })
     }
 
-    /// Creates a new firewall instance.
-    pub fn new(#[cfg(target_os = "linux")] fwmark: u32) -> Result<Self, Error> {
-        Ok(Firewall {
-            inner: imp::Firewall::new(
-                #[cfg(target_os = "linux")]
-                fwmark,
-            )?,
-        })
-    }
-
     /// Applies and starts enforcing the given `FirewallPolicy` Makes sure it is being kept in place
     /// until this method is called again with another policy, or until `reset_policy` is called.
     pub fn apply_policy(&mut self, policy: FirewallPolicy) -> Result<(), Error> {
@@ -439,6 +442,12 @@ impl Firewall {
     pub fn reset_policy(&mut self) -> Result<(), Error> {
         tracing::info!("Resetting firewall policy");
         self.inner.reset_policy()
+    }
+
+    // fixme: exposed as a poor man solution to support nym-split-tunnel
+    #[cfg(target_os = "linux")]
+    pub fn send_and_process(batch: &nftnl::FinalizedBatch) -> Result<(), Error> {
+        imp::Firewall::send_and_process(batch)
     }
 }