+++ nym-vpn-core/crates/nym-split-tunnel/src/lib.rs
@@ -2,7 +2,12 @@
 // Copyright 2026 Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use std::net::{Ipv4Addr, Ipv6Addr};
+#[cfg(target_os = "linux")]
+#[path = "linux/mod.rs"]
+mod imp;
+
+#[cfg(target_os = "linux")]
+pub use imp::*;
 
 #[cfg(target_os = "macos")]
 #[path = "macos/mod.rs"]
@@ -17,6 +22,8 @@ pub use imp::*;
 
 #[cfg(target_os = "windows")]
 pub use imp::{install_driver_service, uninstall_driver_service};
+
+use std::net::{Ipv4Addr, Ipv6Addr};
 
 /// VPN tunnel interface configuration used by split tunneling.
 #[derive(Debug, Clone)]
@@ -33,6 +40,7 @@ pub struct VpnInterface {
 #[derive(Debug, Copy, Clone)]
 pub enum SplitTunnelErrorCause {
     /// Device is offline, split tunneling cannot be used.
+    #[cfg(target_os = "macos")]
     IsOffline,
 
     #[cfg(target_os = "macos")]