+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/mod.rs
@@ -57,9 +57,7 @@ use tokio_util::sync::CancellationToken;
 use nym_dns::DnsConfig;
 #[cfg(not(any(target_os = "android", target_os = "ios")))]
 use nym_firewall::{Firewall, FirewallArguments, InitialFirewallState};
-use nym_gateway_directory::{
-    BlacklistedGateways, Config as GatewayDirectoryConfig, GatewayCacheHandle,
-};
+use nym_gateway_directory::{Config as GatewayDirectoryConfig, GatewayCacheHandle};
 use nym_vpn_lib_types::{
     AccountControllerErrorStateReason, ActionAfterDisconnect, ConnectionData, EntryPoint,
     ErrorStateReason, EstablishConnectionData, EstablishConnectionState, ExitPoint,
@@ -69,6 +67,7 @@ use nym_vpn_lib_types::{
 use crate::{
     GatewayDirectoryError, UserAgent, bandwidth_controller::Error as BandwidthControllerError,
     mixnet::VpnTopologyServiceHandle,
+    tunnel_state_machine::tunnel::gateway_provider::GatewayProvider,
 };
 
 use tunnel::SelectedGateways;
@@ -111,6 +110,10 @@ pub struct TunnelConstants {
     #[cfg(target_os = "linux")]
     /// Firewall mark used for bypassing the tunnel
     pub fwmark: u32,
+
+    /// Tunnel specific routing table, traffic not marked will be routed via this routing table.
+    #[cfg(target_os = "linux")]
+    pub table_id: u32,
 }
 
 impl Default for TunnelConstants {
@@ -123,6 +126,8 @@ impl Default for TunnelConstants {
             ),
             #[cfg(target_os = "linux")]
             fwmark: crate::TUNNEL_FWMARK,
+            #[cfg(target_os = "linux")]
+            table_id: crate::TUNNEL_TABLE_ID,
         }
     }
 }
@@ -600,7 +605,7 @@ pub struct SharedState {
     filtering_resolver: resolver::ResolverHandle,
     #[cfg(not(any(target_os = "android", target_os = "ios")))]
     adblocker: adblocker::AdBlockerTaskHandle,
-    #[cfg(any(windows, target_os = "macos"))]
+    #[cfg(any(target_os = "macos", target_os = "windows"))]
     split_tunnel: nym_split_tunnel::SplitTunnelHandle,
     nym_config: NymConfig,
     tunnel_settings: TunnelSettings,
@@ -613,12 +618,10 @@ pub struct SharedState {
     account_command_tx: AccountCommandSender,
     account_controller_state: AccountStateReceiver,
     statistics_event_sender: StatisticsSender,
-    gateway_cache_handle: GatewayCacheHandle,
+    gateway_provider: GatewayProvider<GatewayCacheHandle>,
     topology_service: VpnTopologyServiceHandle,
     discovery_refresher_command_tx: mpsc::UnboundedSender<DiscoveryRefresherCommand>,
-    wg_keys_db: WireguardKeysDb,
     user_agent: UserAgent,
-    blacklisted_entry_gateways: BlacklistedGateways,
 }
 
 impl SharedState {
@@ -687,7 +690,7 @@ impl SharedState {
             Ok(had_interface != has_interface)
         }
 
-        #[cfg(target_os = "windows")]
+        #[cfg(any(target_os = "linux", target_os = "windows"))]
         {
             Ok(false)
         }
@@ -732,6 +735,17 @@ impl SharedState {
     }
 }
 
+#[cfg(target_os = "linux")]
+pub struct LinuxSplitTunnelConfiguration {
+    /// The cgroup2 used for split tunneling.
+    /// Traffic from processes in this cgroup2 should be allowed outside the tunnel.
+    pub excluded_cgroup2: Option<nym_cgroup::v2::CGroup2>,
+
+    /// The net_cls id of the v1 cgroup used for split tunneling.
+    /// This is used as a fallback to [`Self::excluded_cgroup2`] since old kernels don't support cgroups v2.
+    pub net_cls: Option<u32>,
+}
+
 #[derive(Debug, Clone)]
 pub struct NymConfig {
     pub config_path: Option<PathBuf>,
@@ -753,6 +767,7 @@ pub struct TunnelStateMachine {
     filtering_resolver_handle: JoinHandle<()>,
     #[cfg(not(any(target_os = "android", target_os = "ios")))]
     adblocker_handle: JoinHandle<()>,
+    gateway_provider_handle: JoinHandle<()>,
     shutdown_token: CancellationToken,
 }
 
@@ -772,7 +787,9 @@ impl TunnelStateMachine {
         connectivity_handle: ConnectivityHandle,
         discovery_refresher_command_tx: mpsc::UnboundedSender<DiscoveryRefresherCommand>,
         wg_keys_db: WireguardKeysDb,
-        #[cfg(any(windows, target_os = "macos"))] split_tunnel: nym_split_tunnel::SplitTunnelHandle,
+        #[cfg(any(target_os = "macos", target_os = "windows"))]
+        split_tunnel: nym_split_tunnel::SplitTunnelHandle,
+        #[cfg(target_os = "linux")] split_tunnel_config: LinuxSplitTunnelConfiguration,
         #[cfg(not(any(target_os = "android", target_os = "ios")))] route_handler: RouteHandler,
         #[cfg(target_os = "ios")] tun_provider: Arc<dyn OSTunProvider>,
         #[cfg(target_os = "android")] tun_provider: Arc<dyn AndroidTunProvider>,
@@ -829,6 +846,12 @@ impl TunnelStateMachine {
             initial_state: InitialFirewallState::None,
             #[cfg(target_os = "linux")]
             fwmark: tunnel_constants.fwmark,
+            #[cfg(target_os = "linux")]
+            table_id: tunnel_constants.table_id,
+            #[cfg(target_os = "linux")]
+            excluded_cgroup2: split_tunnel_config.excluded_cgroup2,
+            #[cfg(target_os = "linux")]
+            net_cls: split_tunnel_config.net_cls,
         })
         .map_err(Error::CreateFirewall)?;
 
@@ -839,6 +862,9 @@ impl TunnelStateMachine {
         {
             nym_common::trace_err_chain!(err, "failed to set initial split tunnel paths");
         }
+
+        let (gateway_provider, gateway_provider_handle) =
+            GatewayProvider::new(gateway_cache_handle, wg_keys_db, shutdown_token.clone());
 
         let mut shared_state = SharedState {
             #[cfg(not(any(target_os = "android", target_os = "ios")))]
@@ -863,12 +889,10 @@ impl TunnelStateMachine {
             account_command_tx,
             account_controller_state,
             statistics_event_sender,
-            gateway_cache_handle,
+            gateway_provider,
             topology_service,
             discovery_refresher_command_tx,
-            wg_keys_db,
             user_agent,
-            blacklisted_entry_gateways: BlacklistedGateways::new(),
         };
 
         let (current_state_handler, _) = if shared_state
@@ -895,6 +919,7 @@ impl TunnelStateMachine {
             filtering_resolver_handle,
             #[cfg(not(any(target_os = "android", target_os = "ios")))]
             adblocker_handle,
+            gateway_provider_handle,
             shutdown_token,
         };
 
@@ -947,6 +972,9 @@ impl TunnelStateMachine {
             if let Err(e) = self.adblocker_handle.await {
                 tracing::error!("Failed to join on ad-blocker task: {}", e)
             }
+        }
+        if let Err(e) = self.gateway_provider_handle.await {
+            tracing::error!("Failed to join on gateway provider task: {}", e)
         }
     }
 }
@@ -1053,6 +1081,9 @@ pub enum Error {
     // Temporary until we support `RegistrationResult::Lp()`
     #[error("invalid tunnel type")]
     InvalidTunnelType,
+
+    #[error("gateway provider shut down")]
+    GatewayProviderDown,
 }
 
 impl Error {
@@ -1100,6 +1131,7 @@ impl Error {
             Self::CreateTcpProbe(e) => ErrorStateReason::Internal(e.to_string()),
             Self::ProbeRequiresIPv4Addr => ErrorStateReason::Internal(self.to_string()),
             Self::InvalidTunnelType => ErrorStateReason::Internal(self.to_string()),
+            Self::GatewayProviderDown => ErrorStateReason::Internal(self.to_string()),
         })
     }
 }